requirements.txt
anyio==3.5.0
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
fastapi==0.72.0
h11==0.13.0
httptools==0.3.0
idna==3.3
kazoo==2.5.0
pydantic==1.8.2
pykafka==2.8.0
pymongo==4.0.1
python-dotenv==0.19.2
PyYAML==6.0
requests==2.27.1
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
tabulate==0.8.9
typing_extensions==4.0.1
urllib3==1.26.8
uvicorn==0.17.1
uvloop==0.16.0
watchgod==0.7
websockets==10.1
